pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "data-analysis"
version = "0.1.0"
description = "Personal AI Data Analyst - Automate data analysis with LLMs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.1",
    "langchain-community>=0.3.0",
    "langchain-experimental>=0.0.1",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "matplotlib>=3.8.0",
    "numpy<2.0.0",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "scipy>=1.10.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.30.0",
    "tabulate>=0.9.0",
    "seaborn>=0.13.0",
    "sqlalchemy>=2.0.0",
    "pandasql>=0.7.3",
]

[tool.uv]
managed = true

[tool.hatch.build.targets.wheel]
packages = ["src/data_analysis"]
